stdint and stdbool files made different for kinetis and s08 - need different typedefs.

+++ support/kinetis/ucp_stdint.h
@@ -11,9 +11,9 @@
 #define MSF_STDINT_H
 
 
-typedef unsigned char   uint8_t;
-typedef unsigned int    uint16_t;
-typedef unsigned long   uint32_t;
+typedef unsigned char   	uint8_t;
+typedef unsigned short int	uint16_t;
+typedef unsigned long int   uint32_t;
